requirements.txt
git+https://github.com/Rapptz/discord.py#egg=discord.py
aiohttp
PyNaCl
python-dotenv
